backend/requirements.txt
bcrypt==4.0.1
cloudinary==1.44.1
cryptography==46.0.3
eventlet==0.40.4
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
google-generativeai==0.8.5
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
python-socketio==5.15.0
requests==2.32.5
Werkzeug==3.1.4